+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+* Added no_std compatibility
+
 ## 0.1.0 (2021-04-30)
 
 * First commit